requirements.txt
annotated-types==0.6.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.5
pydantic==2.6.1
pydantic_core==2.16.2
python-decouple==3.8
requests==2.31.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
